Add FakeOutboxWriter for unit testing repositories
Introduced a `FakeOutboxWriter` class in `CategoryRepositoryTests.cs` and `ProductRepositoryTests.cs` to provide a no-op implementation of the `IOutboxWriter` interface for unit testing. Updated the `CreateRepository` methods in both test files to pass the `FakeOutboxWriter` to the respective repository constructors.

Added the `using Common.Utilities.Interfaces.Messaging.Events;` directive to include the `IOutboxWriter` interface. These changes enable isolated testing of repositories with the new `IOutboxWriter` dependency.